feat(NPM): change entry point for module output
We are changing the entry point to ensure cert-verifier-js can consume the right version for node or the browser with rollup.
Since it defaults to using module over main, and that most of the time we want the module entry point on other dependencies, we change hashlink verifier to expose the module build on the import entry point.